+++ src/core/index.ts
@@ -1,6 +1,7 @@
 // auto generated with tsai
 // https://github.com/thomas-crane/tsai
 export * from './client';
+export * from './lib-info';
 export * from './plugin-manager';
 export * from './plugin-module';
 export * from './resource-manager';