+++ src/decorators/index.ts
@@ -2,3 +2,4 @@
 // https://github.com/thomas-crane/tsai
 export * from './hook-packet';
 export * from './plugin';
+export * from './valid-packets';