+++ src/decorators/plugin.ts
@@ -1,9 +1,15 @@
+import 'reflect-metadata';
 import { PluginManager } from './../core/plugin-manager';
-import { IPluginInfo } from './../models/plugin-info';
+import { ILibraryInfo } from './../models/plugin-info';
 
-export function NrPlugin(info: IPluginInfo): (target: new () => object) => any {
-    return (target: new () => object) => {
-        PluginManager.addPlugin(info, target);
-        return target;
+export function Library(libInfo: ILibraryInfo): ClassDecorator {
+    return (target: any) => {
+        const params = Reflect.getMetadata('design:paramtypes', target) || [];
+        const dependencies = params.map((type: any) => type.name);
+        PluginManager.loadLibrary({
+            info: libInfo,
+            target,
+            dependencies
+        });
     };
 }