+++ src/plugins/hello-plugin.ts
@@ -1,13 +1,13 @@
-// These 5 imports are essential for any plugin.
-import { NrPlugin, HookPacket, Packet, PacketType, Client } from './../core';
+// These imports are essential for any plugin.
+import { Client, Library, PacketHook } from './../core';
 
 // These 2 imports are only required because this plugin deals with text packets.
 import { TextPacket } from './../networking/packets/incoming';
 import { PlayerTextPacket } from './../networking/packets/outgoing';
 
-// The NrPlugin decorator gives nrelay some information about
+// The Library decorator gives nrelay some information about
 // your plugin. If it is not present, nrelay won't load the plugin.
-@NrPlugin({
+@Library({
     name: 'Hello Plugin',
     author: 'tcrane'
 })
@@ -15,7 +15,7 @@ class HelloPlugin {
 
     // The HookPacket decorator will cause the method to be called
     // whenever a packet with the specified packet type is recieved.
-    @HookPacket(PacketType.TEXT)
+    @PacketHook()
     // Any method with a HookPacket decorator should always have
     // the method signature (client: Client, packet: Packet).
     onText(client: Client, textPacket: TextPacket): void {